refactor: 重构并中心化应用的日志系统
对应用的日志系统进行了结构性重构，以实现单一配置源和更清晰的模块职责。
主要变更：
- 将 GuiLoggingHandler 提取到独立的 log_utils.py 文件中，使其成为一个独立的工具模块。
- 清理了业务模块 (如 bili_sender.py) 中所有分散的、局部的日志配置代码。
