+++ monitor_tab.py
@@ -25,7 +25,7 @@ class MonitorTab(ttk.Frame):
         super().__init__(parent, padding=15)
         self.model = model
         self.app = app
-        self.logger = logging.getLogger("monitor_tab")
+        self.logger = logging.getLogger("MonitorTab")
         # 配置主Frame的列和行权重
         self.columnconfigure(0, weight=1)
         self.rowconfigure(2, weight=1)
